+++ workflow_engine/project_apps/engine/job_dependency.py
@@ -5,7 +5,8 @@ from celery.exceptions import SoftTimeLimitExceeded
 
 from project_apps.constants import JOB_STATUS_WAITING
 from project_apps.models.cache import Cache
-from project_apps.engine.job_execute import job_execute
+from project_apps.engine.tasks_manager import job_execute
+
 
 @shared_task
 def job_dependency(workflow_uuid, history_uuid):
@@ -20,7 +21,7 @@ def job_dependency(workflow_uuid, history_uuid):
 
         for job_data in job_list:
             if job_data['depends_count'] == 0 and job_data['result'] == JOB_STATUS_WAITING:
-                job_execute.apply_async(args=[workflow_uuid, history_uuid, job_data['uuid']])
+                job_execute(workflow_uuid, history_uuid, job_data['uuid'])
 
         return {'status': 'success', 'message': 'Jobs execution started successfully'}
